+++ src/core/settings/formula.ts
@@ -14,10 +14,10 @@ export type FormulaSettings = {
 export const DEFAULT_FORMULA_SETTINGS: FormulaSettings = {
     clickCopyMarkdown: true,
     assetActions: {
-        copyPng: true,
-        copySvg: true,
-        copyMathml: true,
-        savePng: true,
-        saveSvg: true,
+        copyPng: false,
+        copySvg: false,
+        copyMathml: false,
+        savePng: false,
+        saveSvg: false,
     },
 };